requirements.txt
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
idna==2.10
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pyOpenSSL==19.1.0
requests==2.24.0
six==1.16.0
toml==0.10.2
urllib3==1.25.11
